requirements.txt
django==4.1.7
pillow==9.4.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-recaptcha==3.0.0